add column stats, cardinality-based tree scoring, update experiment.
- collect per-(label, property) HLL/min/max/count in FlatGraph (NDV via
  HyperLogLog P=6, similar to Kuzu/DuckDB)
- live-update stats on random_insert so the catalog stays consistent
- replace the heuristic edge score (1-5) with edge cardinality estimated
  from stats; spanning-tree builder now keeps low-cardinality edges and
  drops large-cardinality edges from cycles
- bump CandidateTree.total_score from u32 to u64 (cardinality sums) and
  reverse Ord so BinaryHeap pops the tightest tree first
- new run_insert_impact.sh + deploy_and_run.md for the 5-round insert
  experiment across L1-L6 patterns

+++ minigu/core/src/procedures/gcard_query/mod.rs
@@ -263,9 +263,9 @@ pub fn build_procedure() -> Procedure {
                 let estimate_start = Instant::now();
                 let total_count = abstract_graphs_with_scores.len();
                 let mut min_nonzero_es = f64::INFINITY;
-                let mut score_of_min_es: Option<u32> = None;
+                let mut score_of_min_es: Option<u64> = None;
                 let mut index_of_min_es: Option<usize> = None;
-                let mut max_score: u32 = 0;
+                let mut max_score: u64 = 0;
                 #[allow(dead_code)]
                 let mut min_es_abstract_graph: Option<AbstractGraph> = None;
                 for (idx, (mut abs, score)) in abstract_graphs_with_scores.into_iter().enumerate() {
